Revert permalink changes for files
