+++ src/Uuid/PageUuid.php
@@ -54,4 +54,25 @@ class PageUuid extends ModelUuid
 			yield from static::index($page);
 		}
 	}
+
+	/**
+	 * Returns permalink url
+	 */
+	public function url(): string
+	{
+		// make sure UUID is cached because the permalink
+		// route only looks up UUIDs from cache
+		if ($this->isCached() === false) {
+			$this->populate();
+		}
+
+		$kirby = App::instance();
+		$url   = $kirby->url();
+
+		if ($language = $kirby->language('current')) {
+			$url .= '/' . $language->code();
+		}
+
+		return $url . '/@/' . static::TYPE . '/' . $this->id();
+	}
 }
